requirements.txt
langchain==0.0.336
langchain_experimental
streamlit
openai==1.3.0
tiktoken
tqdm
python-dotenv
sentence_transformers
tabulate
numpy